--- Common/DTO/AdminEmployeeDTO.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Common.DTO
-{
-    public class AdminEmployeeDTO
-    {
-        public string EmployeeCode { get; set; }
-        public string Name { get; set; }
-        public string Designation { get; set; }
-    }
-}
