+++ Skillset-BLL/Services/IEmployeeServices.cs
@@ -70,5 +70,19 @@ namespace Skillset_BLL.Services
         string GetRoleName(string id);
        
         EmployeeDTO GetProfile(string id);
+
+        /// <summary>
+        /// Get Skill details of an employee from table Skillrating
+        /// </summary>
+        /// <param name="employeeCode"></param>
+        /// <returns></returns>
+        IEnumerable<AdminSkillDTO> GetSkillDetails(string employeeCode);
+
+        /// <summary>
+        /// Finding employee name from table Employee
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        string GetEmployeeName(string id);
     }
 }
