+++ Skillset-BLL/Skillset-BLL.csproj
@@ -44,13 +44,11 @@
     <Compile Include="obj\Debug\TemporaryGeneratedFile_5937a670-0e60-4077-877b-f7221da3dda1.cs" />
     <Compile Include="obj\Debug\TemporaryGeneratedFile_E7A71F73-0F8D-4B9B-B56E-8E70B10BC5D3.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\AdminEmployeeSkillService.cs" />
     <Compile Include="Services\EmployeeServices.cs" />
     <Compile Include="Services\IEmployeeServices.cs" />
     <Compile Include="Services\ILoginService.cs" />
     <Compile Include="Services\LoginService.cs" />
     <Compile Include="Services\IReportingStaffExtensions.cs" />
-    <Compile Include="Services\IAdminEmployeeSkillService.cs" />
     <Compile Include="Services\ISkillRatingService.cs" />
     <Compile Include="Services\ISkillService.cs" />
     <Compile Include="Services\ReportingStaffExtensions.cs" />
