+++ Skillset-DAL/Repositories/IEmployeeRepository.cs
@@ -113,5 +113,39 @@ namespace Skillset_DAL.Repositories
         /// </summary>
         /// <returns></returns>
         Employee GetProfile(string id);
+
+        /// <summary>
+        /// Get Skill details of an employee from table Skillrating
+        /// </summary>
+        /// <param name="employeeCode"></param>
+        /// <returns></returns>
+        List<SkillRating> GetSkillDetails(string employeeCode);
+
+        /// <summary>
+        /// Finding skill name from table Skill
+        /// </summary>
+        /// <param name="skillId"></param>
+        /// <returns></returns>
+        string FindSkillName(int skillId);
+
+        /// <summary>
+        /// Finding skill value from table Rating
+        /// </summary>
+        /// <param name="ratingId"></param>
+        /// <returns></returns>
+        int FindSkillValue(int ratingId);
+
+        /// <summary>
+        /// Finding employee name from table Employee
+        /// </summary>
+        /// <param name="employeeCode"></param>
+        /// <returns></returns>
+        string FindEmployeeName(string employeeCode);
+        /// <summary>
+        ///  Finding note of a rating value from table Rating
+        /// </summary>
+        /// <param name="ratingId"></param>
+        /// <returns></returns>
+        string FindRatingNote(int ratingId);
     }
 }
