--- Skillset-PL/Views/AdministratorView/Index.cshtml
@@ -1,93 +1,0 @@
-﻿@model IPagedList<Skillset_PL.ViewModels.AdministratorEmployeeViewModel>
-@using PagedList.Mvc;
-@using PagedList;
-@{
-    Layout = "~/Views/Shared/_AdminDashboardLayout.cshtml";
-    ViewBag.Title = "View Employee Skills";
-}
-<div class="row first">
-    <h2>Employee Skill List</h2>
-    <br />
-    <div class="panel-group">
-        <div class="panel panel-success">
-            <div class="panel-heading" style="display:block;text-align:center"><b>Search</b></div>
-            <div class="panel-body">
-                <div class="form-group">
-                    <div class="col-md-5">
-                        <b> Search By Name/Code/Designation</b>
-                    </div>
-
-                    @using (Html.BeginForm("Index", "AdministratorView", FormMethod.Get))
-                    {
-                        <div class="col-md-3">
-                            @Html.Editor("search", null, new { htmlAttributes = new { @class = "form-control" } })
-                            &nbsp; &nbsp;
-                        </div>
-                        <div class="col-md-4">
-                            <input type="submit" name="submit" value="Search" class="btn btn-success" />
-                            &nbsp;
-                            @Html.ActionLink("Reset", "Index", null, htmlAttributes: new { @class = "btn btn-success" })
-                        </div>
-                    }
-                </div>
-            </div>
-        </div>
-    </div>
-
-    @if (Model.Count() > 0)
-    {
-        
-        <br />
-        <label class="control-label alert-success well-sm" style="display:block;text-align:center">Employee Details</label>
-        <table class="table table-bordered table-hover table-responsive">
-            <thead>
-                <tr>
-                    <th>
-                        @Html.DisplayName("Employee Code")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Name")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Designation")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Action")
-                    </th>
-                </tr>
-            </thead>
-            <tbody>
-                @foreach (var item in Model)
-                {
-                    <tr>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.EmployeeCode)
-                        </td>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.Name)
-                        </td>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.Designation)
-                        </td>
-                        <td>
-                            @Html.ActionLink("Skills", "Skills", new { id = item.EmployeeCode }, htmlAttributes: new { @class = "btn btn-primary" })
-                        </td>
-                    </tr>
-                }
-            </tbody>
-        </table>
-    }
-    else
-    {
-        <div class="alert alert-warning fade in">
-            <a href="#" class="close" data-dismiss="alert">&times;</a>
-
-            <strong>No Employees!</strong> There are no employees to display.
-
-        </div>
-    }
-    Page @(Model.PageCount < Model.PageNumber ? 0 : Model.PageNumber) of @Model.PageCount
-    @Html.PagedListPager(Model, pageNo => Url.Action("Index", new { search = ViewBag.search,pageNo }))
-</div>
-
-
