--- Skillset-PL/Views/AdministratorView/Skills.cshtml
@@ -1,79 +1,0 @@
-﻿@model IEnumerable<Skillset_PL.ViewModels.AdministratorSkillViewModel>
-
-@{
-    Layout = "~/Views/Shared/_AdminDashboardLayout.cshtml";
-    ViewBag.Title = "Skills";
-}
-<link href="~/Content/Site.css" rel="stylesheet" type="text/css" />
-<div class="row first">
-    <h2>Skills</h2>
-
-    <label class="control-label alert-success well-sm" style="display:block;text-align:center">Skills</label>
-    <label class="control-label col-sm-6" style="float:left;display:block;text-align:center">Employee Code : @(ViewData["employeecode"]) </label>
-    <label class="control-label col-sm-6" style="float:left;display:block;text-align:left;">  Employee Name : @(ViewData["employeename"]) </label>
-    <br />
-    <br />
-    @if (Model.Count() > 0)
-    {
-        <table class="table table-bordered table-hover table-responsive">
-            <thead>
-                <tr>
-                    <th>
-                        @Html.DisplayName("Skill")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Rating")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Rated Date")
-                    </th>
-                    <th>
-                        @Html.DisplayName("Note")
-                    </th>
-                </tr>
-            </thead>
-            <tbody>
-                @foreach (var item in Model)
-                {
-                    <tr>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.SkillName)
-                        </td>
-                        <td>
-                            <div data-tooltip="@(item.RatingNote)" data-tooltip-position="bottom">
-                                @for (var i = 1; i <= item.SkillValue; i++)
-                                {
-                                    <span class="starGlowTable"></span>
-                                }
-                                @for (var i = (item.SkillValue + 1); i <= 5; i++)
-                                {
-                                    <span class="starFadeTable"></span>
-                                }
-                            </div>
-
-                        </td>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.RatingDate)
-                        </td>
-                        <td>
-                            @Html.DisplayFor(modelItem => item.Note)
-                        </td>
-                    </tr>
-                }
-            </tbody>
-        </table>
-    }
-    else
-    {
-        <br />
-        <div class="alert alert-warning fade in">
-
-            <strong>No Skill Ratings!</strong> There is no skill ratings available.
-        </div>
-    }
-    <br />
-    <br />
-    <p>
-        @Html.ActionLink("Back to Employee List", "Index", null, htmlAttributes: new { @class = "btn btn-primary" })
-    </p>
-    </div>
