+++ Skillset-PL/Views/Shared/_AdminDashboardLayout.cshtml
@@ -53,11 +53,7 @@
                     </a>
                 </li>
 
-                <li>
-                    <a href="~/AdministratorView/Index">
-                        <i class="fa fa-eye" style="font-size:20px"></i> View Employee Skills
-                    </a>
-                </li>
+               
             </ul>
         </div>
     </div>
